requirements.txt
Django==1.11.1
django-allauth==0.32.0
django-bootstrap3==8.2.2
django-filter==1.0.4
django-material==0.13.1
django-widget-tweaks==1.4.1

